WIP: References
This commit does not compile and does not work correctly at all.

It is an unfinished attempt to resolve only one type of references - signal references to methods within a single file. There is a huge amount of TODOs and NotImplementExceptions. A more generic approach has to be taken to properly implement this feature.