Adjust .tscn grammar for typed collections
With the addition of typed Arrays and Dictionaries in Godot 4.x the
grammar for resource files has changed slightly.  In addition, it is
possible for dictionaries in resources to have arbitrary key types (as
shown in the included test file) so I tweaked the grammar to account for
that as well.

I generated the parser code using IntelliJ and backed out any
formatting-only changes to reduce the footprint of this PR.

The change to `GodotResourcePathCodeCompletion.cs` was necessary to make
Rider run; I'm not exactly sure why, and I'm happy to revert it.